WIP: AaveOracle change to support USD oracles
In Agave, our oracles are based in USD. However, there is no native crypto asset
which is *exactly* pegged to USD.
Internally, Aave prices things in terms of the native token, and assumes
that all oracles are priced in the native token (ETH). But on sidechains
with different native tokens (e.g. xDai), it may be difficult to get
a reliable oracle priced in terms of the native token.

So, in getAssetPrice we now take the price of `WETH` in USD. `WETH` here
refers to "the wrapped version of the native token".
(Technically, this doesn't have to be USD; it just must be the same
price basis as all of the other assets are using.)
Then we divide the returned asset price by this value, and now we have
the value of the asset in terms of the native token.